comparison function (INCOMPLETE)
this is a hard coded version of the function that will be used to perform the comparison tab. I wrote it to be hard coded for now because we are not ready to impliment this in the shiny syntax yet, and that will require this function to work with data that will be generated from others. So for now, we will leave that work to be done when we write the shiny code for this tab.